app/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
ecs-logging==2.0.2
dynaconf==3.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
filelock==3.13.1
fsspec==2024.2.0
gdown==5.2.0
grpcio==1.67.1
h11==0.14.0
idna==3.10
Jinja2==3.1.3
MarkupSafe==2.1.5
milvus-lite==2.4.10
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.2.1
numpy==1.26.3
pandas==2.2.3
pillow==10.2.0
protobuf==5.29.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymilvus==2.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
